@@ -1,14 +1,15 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import { createMemoryHistory, createRouter } from 'vue-router'
+//import Router from 'vue-router'
 import about from './views/about.vue'
 import contact from './views/contact.vue'
 import skills from './views/skills.vue'
 import portfolio from './views/portfolio.vue'
 
-Vue.use(Router)
 
-export default new Router({
+
+export const router = createRouter({
     mode: 'hash',
+    history: createMemoryHistory(),
     routes: [
         {
             path: '/',
